trie: drop subdomains from final trie
Treat every blocklist entry as if it was a wildcard entry. This means, there
isn't a need to store a subdomain of domain that's already in the trie against
that particular blocklist. That is, if a blocklist contains www.example.com
and example.com, then it is enough to store just example.com.

For now, enable it for both u8 and u6 codec, but in prod, this trimming down
should only be enabled for codec u8.